+++ game/Airside/Assets/Airside/Simulation/GroundTraffic.cs
@@ -213,7 +213,7 @@ namespace Airside.Simulation
         public static GroundLeg LegFor(FleetAircraft aircraft, FleetGroundLeg leg) => leg switch
         {
             FleetGroundLeg.TaxiOut => AdelaideGround.TaxiOut(aircraft.DepartureStand, aircraft.Type, aircraft.AssignedRunway),
-            FleetGroundLeg.Lineup => AdelaideGround.LineupFor(aircraft.AssignedRunway),
+            FleetGroundLeg.Lineup => AdelaideGround.LineupFor(aircraft.AssignedRunway, aircraft.Type),
             FleetGroundLeg.Vacate => AdelaideGround.VacateFor(aircraft.Type, aircraft.AssignedRunway),
             _ => AdelaideGround.TaxiIn(aircraft.Stand, aircraft.Type, aircraft.AssignedRunway)
         };